optimized data types

+++ src/com/softserve/edu03/Employee.java
@@ -2,20 +2,20 @@ package com.softserve.edu03;
 
 public class Employee {
     private String name;
-    private double rate;
+    private float rate;
     private int hours;
-    private static double totalSum;
+    private static float totalSum;
 
     public Employee() {
 
     }
 
-    public Employee (String name, double rate){
+    public Employee (String name, float rate){
         this.name = name;
         this.rate = rate;
     }
 
-    public Employee (String name, double rate, int hours){
+    public Employee (String name, float rate, int hours){
         this.name = name;
         this.rate = rate;
         this.hours = hours;
@@ -29,11 +29,11 @@ public class Employee {
         return this.name;
     }
 
-    public void setRate(double rate) {
+    public void setRate(float rate) {
         this.rate = rate;
     }
 
-    public double getRate() {
+    public float getRate() {
         return this.rate;
     }
 
@@ -45,7 +45,7 @@ public class Employee {
         return this.hours;
     }
 
-    public double getSalary() {
+    public float getSalary() {
         totalSum += this.rate * this.hours;
         return this.rate * this.hours;
     }
@@ -56,19 +56,19 @@ public class Employee {
                 " he worked " + getHours() + " hours.";
     }
 
-    public double changeRate(double rate){
+    public float changeRate(float rate){
         totalSum -= (this.rate * this.hours);
         this.rate = rate;
         getSalary();
         return this.rate;
     }
 
-    public double getBonuses() {
+    public float getBonuses() {
         totalSum += this.rate * this.hours / 10;
         return this.rate * this.hours / 10;
     }
 
-    public static double getTotalSum() {
+    public static float getTotalSum() {
         return totalSum;
     }
 }
